// driver-front/projects/app-company/src/app/models/PlannigTrip.ts
export class PlanningTrip {

    id: number;
    departCity: string;
    arrivalCity: string;
    departDate: string;
    arrivalDate: string;
    departHour: string;
    arrivalHour: string;
    type: string;
    tripPrice: number;
    companyId: number;
    reservationNB: number;

    constructor() {}

}